+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 8.0.0 (IN PROGRESS)
 
 * Increment `react-router` to `^5.2`.
+* Upgraded `react-overlays` dependency to the latest version. Refs STCOM-650.
 * Add a utility list of language names & helper functions. Refs UIIN-829.
 * Added `<TextLink>`-component. Refs STCOM-699.
 * Extend `ConfirmationModal` interactor. STCOM-710.